Separate controller test class from configuration
